# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PDF Exporter REST API",
    "version": "v1"
  },
  "paths": {
...
    "/api/settings/cover-page/names/{name}/images": {
      "delete": {
        "operationId": "deleteImages",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "schema": {
              "default": "",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Images deleted successfully"
          }
        },
        "summary": "Deletes images in SVN linked to specified cover page within specified scope (global or certain project)",
        "tags": [
          "Settings"
        ]
      }
    },
...
    "/api/settings/cover-page/templates/{template}": {
      "post": {
        "operationId": "persistCoverPageTemplate",
        "parameters": [
          {
            "in": "path",
            "name": "template",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Template persisted successfully"
          }
        },
        "summary": "Persist content of cover page predefined template",
        "tags": [
          "Settings"
        ]
      }
    },
    "/api/settings/localization/names/{name}/download": {
      "get": {
        "operationId": "downloadTranslations",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "language",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "revision",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Localization file downloaded successfully"
          }
        },
        "summary": "Downloads localization values by name of localization settings",
        "tags": [
          "Settings"
        ]
      }
    },
    "/api/settings/localization/upload": {
      "post": {
        "operationId": "uploadTranslations",
        "parameters": [
          {
            "in": "query",
            "name": "language",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/FormDataBodyPart"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Localization file uploaded successfully"
          }
        },
        "summary": "Uploads localization values",
        "tags": [
          "Settings"
        ]
      }
    },
    "/api/settings/style-package/suitable-names": {
      "post": {
        "operationId": "getSuitableStylePackageNames",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/DocIdentifier"
                },
                "type": "array"
              }
            }
          },
          "description": "List of document identifiers",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully retrieved the list of style package names"
          }
        },
        "summary": "Get list of style packages suitable for the specified list of documents (sorted by weight)",
        "tags": [
          "Settings"
        ]
      }
    },
    "/api/settings/style-package/weights": {
      "get": {
        "operationId": "getStylePackageWeights",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the list of style package names with the weight information"
          }
        },
        "summary": "Get full list of available style packages for the specific scope with the weight information",
        "tags": [
          "Settings"
        ]
      },
      "post": {
        "operationId": "updateStylePackageWeights",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/StylePackageWeightInfo"
                },
                "type": "array"
              }
            }
          },
          "description": "Style packages list with the weight information",
          "required": true
        },
        "responses": {
          "default": {
            "content": {
              "*/*": {}
            },
            "description": "default response"
          }
        },
        "summary": "Update weight information for the provided style packages",
        "tags": [
          "Settings"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BodyPart": {
        "properties": {
          "contentDisposition": {
            "$ref": "#/components/schemas/ContentDisposition"
          },
          "entity": {
            "type": "object"
          },
          "headers": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "properties": {
              "empty": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "mediaType": {
            "$ref": "#/components/schemas/MediaType"
          },
          "messageBodyWorkers": {
            "$ref": "#/components/schemas/MessageBodyWorkers"
          },
          "parameterizedHeaders": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ParameterizedHeader"
              },
              "type": "array"
            },
            "properties": {
              "empty": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "parent": {
            "$ref": "#/components/schemas/MultiPart"
          },
          "providers": {
            "$ref": "#/components/schemas/Providers"
          }
        },
        "type": "object"
      },
...
      "ContentDisposition": {
        "properties": {
          "creationDate": {
            "format": "date-time",
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "modificationDate": {
            "format": "date-time",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "readDate": {
            "format": "date-time",
            "type": "string"
          },
          "size": {
            "format": "int64",
            "type": "integer"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "DocIdentifier": {
        "description": "Unique document identifier data",
        "properties": {
          "documentName": {
            "description": "Document name",
            "type": "string"
          },
          "projectId": {
            "description": "Project ID",
            "type": "string"
          },
          "spaceId": {
            "description": "Space ID",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "FormDataBodyPart": {
        "properties": {
          "content": {
            "type": "object"
          },
          "contentDisposition": {
            "$ref": "#/components/schemas/ContentDisposition"
          },
          "entity": {
            "type": "object"
          },
          "fileName": {
            "type": "string"
          },
          "formDataContentDisposition": {
            "$ref": "#/components/schemas/FormDataContentDisposition"
          },
          "headers": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "properties": {
              "empty": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "mediaType": {
            "$ref": "#/components/schemas/MediaType"
          },
          "messageBodyWorkers": {
            "$ref": "#/components/schemas/MessageBodyWorkers"
          },
          "name": {
            "type": "string"
          },
          "parameterizedHeaders": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ParameterizedHeader"
              },
              "type": "array"
            },
            "properties": {
              "empty": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "parent": {
            "$ref": "#/components/schemas/MultiPart"
          },
          "providers": {
            "$ref": "#/components/schemas/Providers"
          },
          "simple": {
            "type": "boolean"
          },
          "value": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "FormDataContentDisposition": {
        "properties": {
          "creationDate": {
            "format": "date-time",
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "modificationDate": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "parameters": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "readDate": {
            "format": "date-time",
            "type": "string"
          },
          "size": {
            "format": "int64",
            "type": "integer"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "MediaType": {
        "properties": {
          "parameters": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "subtype": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "wildcardSubtype": {
            "type": "boolean"
          },
          "wildcardType": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "MessageBodyWorkers": {
        "type": "object"
      },
      "MultiPart": {
        "properties": {
          "bodyParts": {
            "items": {
              "$ref": "#/components/schemas/BodyPart"
            },
            "type": "array"
          },
          "contentDisposition": {
            "$ref": "#/components/schemas/ContentDisposition"
          },
          "entity": {
            "type": "object"
          },
          "headers": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "properties": {
              "empty": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "mediaType": {
            "$ref": "#/components/schemas/MediaType"
          },
          "messageBodyWorkers": {
            "$ref": "#/components/schemas/MessageBodyWorkers"
          },
          "parameterizedHeaders": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ParameterizedHeader"
              },
              "type": "array"
            },
            "properties": {
              "empty": {
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "parent": {
            "$ref": "#/components/schemas/MultiPart"
          },
          "providers": {
            "$ref": "#/components/schemas/Providers"
          }
        },
        "type": "object"
      },
...
      "ParameterizedHeader": {
        "properties": {
          "parameters": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "value": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "Providers": {
        "type": "object"
      },
...
      "StylePackageWeightInfo": {
        "properties": {
          "name": {
            "description": "The name of the setting",
            "type": "string"
          },
          "scope": {
            "description": "The scope of the setting",
            "type": "string"
          },
          "weight": {
            "description": "The weight of the setting",
            "format": "float",
            "type": "number"
          }
        },
        "type": "object"
      },
...
    }
  },
...
}